+++ packages/saas-ui-date-picker/src/use-calendar.ts
@@ -1,14 +1,27 @@
 import { AriaCalendarProps } from '@react-aria/calendar'
 import { useCalendar as useAriaCalendar } from '@react-aria/calendar'
 import { useCalendarState } from '@react-stately/calendar'
-import { createCalendar } from '@internationalized/date'
 import { useRef, useState, useMemo } from 'react'
 import { useDatePickerContext } from './date-picker-context'
 import { DateValue } from './types'
 import { CalendarAction } from './calendar-context'
+import { GregorianCalendar } from '@internationalized/date'
+
+export const defaultCreateCalendar = (name: string) => {
+  switch (name) {
+    case 'gregory':
+      return new GregorianCalendar()
+    default:
+      throw new Error(`Unsupported calendar ${name}`)
+  }
+}
 
 export const useCalendar = (props: AriaCalendarProps<DateValue>) => {
-  const { locale, calendarProps: contextCalendarProps } = useDatePickerContext()
+  const {
+    locale,
+    calendarProps: contextCalendarProps,
+    createCalendar = defaultCreateCalendar,
+  } = useDatePickerContext()
 
   const [action, setAction] = useState<CalendarAction>('calendar')